nutrikid-backend/requirements.txt
fastapi
uvicorn
torch
transformers
sentence-transformers
faiss-cpu
accelerate
huggingface_hub
numpy
pydantic
